feat: api может решать 9х9 судоку
Реализовал простой api c одним энпоинтом:
- `/solve` - решает судоку

Изменил структуру `solver-service`. Добавил пакеты:
- `handlers` - обработчики запросов
- `solver` - решение судоку
- `models` - модели запросов
