+++ solver_service/models/sudoku.go
@@ -1,0 +1,5 @@
+package models
+
+type SudokuRequest struct {
+	Matrix [][]int `json:"matrix"`
+}
